# api/openapi.yaml
openapi: 3.1.0
info:
    title: PP Prime Construction API
    version: 0.1.0
    description: |
        สัญญา API ระหว่างหน้าบ้าน (Angular) และหลังบ้านของระบบบริหารงานก่อสร้าง PP Prime Construction

        หลักการ
        - เงินเป็นบาท (number) ทศนิยมได้ 2 ตำแหน่ง
        - วันที่ตามแผนใช้ `date` (YYYY-MM-DD) เวลาที่เกิดเหตุการณ์ใช้ `date-time` (ISO 8601 พร้อมเขตเวลา)
        - สถานะ/ประเภทเป็นรหัสภาษาอังกฤษ หน้าบ้านแปลเป็นข้อความไทยเอง
        - ข้อผิดพลาดตอบเป็น `application/problem+json` (RFC 9457)
        - การคำนวณทางธุรกิจ (ระดับผู้อนุมัติ กำไร-ขาดทุน ไทม์ไลน์ สถานะงวด) ทำที่หลังบ้าน
        - ทุกการเปลี่ยนแปลงข้อมูล หลังบ้านต้องบันทึก Audit Log เอง (หน้าบ้านไม่มี API เขียน Audit Log)
        - สิทธิ์การเข้าถึงตรวจที่หลังบ้านตามบทบาทใน `GET /roles`; ข้อมูลที่บทบาทนั้นไม่มีสิทธิ์เห็นต้องไม่ถูกส่งมา
servers:
    - url: /api
      description: ค่าเริ่มต้น (หน้าบ้านตั้ง apiBaseUrl ได้)
security:
    - bearerAuth: []
...
paths:
...
    /projects/{code}/assignments:
        parameters:
            - $ref: '#/components/parameters/ProjectCode'
        get:
            tags: [subcontractors]
            summary: ผู้รับผิดชอบ (บุคลากร) และผู้รับเหมาช่วงของโครงการ
            description: ความคืบหน้า/สถานะของผู้รับเหมาช่วง หลังบ้านคำนวณจากขั้นตอนที่มอบหมายในไทม์ไลน์; contractValue ส่งเฉพาะผู้มีสิทธิ์ project.manage
            operationId: getProjectAssignments
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/ProjectAssignments' }
                '404': { $ref: '#/components/responses/NotFound' }
                '409': { $ref: '#/components/responses/Conflict' }
    /projects/{code}/staff:
        parameters:
            - $ref: '#/components/parameters/ProjectCode'
        post:
            tags: [subcontractors]
            summary: มอบหมายบุคลากรเป็นผู้รับผิดชอบในโครงการ (ต้องมีสิทธิ์ project.manage)
            description: ผู้จัดการโครงการคนแรกในรายการคือ responsibleName ของโครงการ (หลังบ้านอัปเดตให้); คนเดิมซ้ำในบทบาทเดิมตอบ 409
            operationId: addProjectStaff
            requestBody:
                required: true
                content:
                    application/json:
                        schema: { $ref: '#/components/schemas/StaffAssignmentInput' }
            responses:
                '201':
                    description: เพิ่มแล้ว
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/StaffAssignment' }
                '403': { $ref: '#/components/responses/Forbidden' }
                '404': { $ref: '#/components/responses/NotFound' }
                '409': { $ref: '#/components/responses/Conflict' }
                '422': { $ref: '#/components/responses/ValidationError' }
    /projects/{code}/staff/{assignmentId}:
        parameters:
            - $ref: '#/components/parameters/ProjectCode'
            - { name: assignmentId, in: path, required: true, schema: { type: string } }
        put:
            tags: [subcontractors]
            summary: แก้บทบาท/หมายเหตุของผู้รับผิดชอบ
            operationId: updateProjectStaff
            requestBody:
                required: true
                content:
                    application/json:
                        schema: { $ref: '#/components/schemas/StaffAssignmentInput' }
            responses:
                '200':
                    description: บันทึกแล้ว
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/StaffAssignment' }
                '403': { $ref: '#/components/responses/Forbidden' }
                '404': { $ref: '#/components/responses/NotFound' }
                '409': { $ref: '#/components/responses/Conflict' }
                '422': { $ref: '#/components/responses/ValidationError' }
        delete:
            tags: [subcontractors]
            summary: นำผู้รับผิดชอบออกจากโครงการ — โครงการต้องมีผู้จัดการโครงการอย่างน้อย 1 คน (409)
            operationId: removeProjectStaff
            responses:
                '204': { description: นำออกแล้ว }
                '403': { $ref: '#/components/responses/Forbidden' }
                '404': { $ref: '#/components/responses/NotFound' }
                '409': { $ref: '#/components/responses/Conflict' }
    /projects/{code}/subcontractors:
        parameters:
            - $ref: '#/components/parameters/ProjectCode'
        post:
            tags: [subcontractors]
            summary: มอบหมายผู้รับเหมาช่วงตามขั้นตอนงาน (ต้องมีสิทธิ์ project.manage)
            description: ต้องตั้งค่างานก่อสร้างแล้ว (ใช้ขั้นตอนจากไทม์ไลน์); ผู้รับเหมาที่ปิดใช้งานมอบหมายไม่ได้ (422); ผู้รับเหมารายเดิมซ้ำในโครงการเดียวกันตอบ 409 (ให้แก้ไขรายการเดิม)
            operationId: addProjectSubcontractor
            requestBody:
                required: true
                content:
                    application/json:
                        schema: { $ref: '#/components/schemas/SubcontractorAssignmentInput' }
            responses:
                '201':
                    description: มอบหมายแล้ว
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/SubcontractorAssignment' }
                '403': { $ref: '#/components/responses/Forbidden' }
                '404': { $ref: '#/components/responses/NotFound' }
                '409': { $ref: '#/components/responses/Conflict' }
                '422': { $ref: '#/components/responses/ValidationError' }
    /projects/{code}/subcontractors/{assignmentId}:
        parameters:
            - $ref: '#/components/parameters/ProjectCode'
            - { name: assignmentId, in: path, required: true, schema: { type: string } }
        put:
            tags: [subcontractors]
            summary: แก้ขอบเขตงาน ขั้นตอน มูลค่าจ้าง
            operationId: updateProjectSubcontractor
            requestBody:
                required: true
                content:
                    application/json:
                        schema: { $ref: '#/components/schemas/SubcontractorAssignmentInput' }
            responses:
                '200':
                    description: บันทึกแล้ว
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/SubcontractorAssignment' }
                '403': { $ref: '#/components/responses/Forbidden' }
                '404': { $ref: '#/components/responses/NotFound' }
                '409': { $ref: '#/components/responses/Conflict' }
                '422': { $ref: '#/components/responses/ValidationError' }
        delete:
            tags: [subcontractors]
            summary: ยกเลิกการมอบหมาย — ลบไม่ได้เมื่อขั้นตอนที่รับผิดชอบเริ่มงานแล้ว (409) ให้แก้ไขแทน
            operationId: removeProjectSubcontractor
            responses:
                '204': { description: ยกเลิกแล้ว }
                '403': { $ref: '#/components/responses/Forbidden' }
                '404': { $ref: '#/components/responses/NotFound' }
                '409': { $ref: '#/components/responses/Conflict' }
    /subcontractors:
        get:
            tags: [subcontractors]
            summary: ทะเบียนผู้รับเหมาช่วง
            operationId: listSubcontractors
            parameters:
                - { name: q, in: query, description: ค้นหาชื่อ รหัส ผู้ติดต่อ, schema: { type: string } }
                - { name: trade, in: query, schema: { $ref: '#/components/schemas/SubcontractorTrade' } }
                - { name: status, in: query, schema: { type: string, enum: [active, inactive] } }
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: array
                                items: { $ref: '#/components/schemas/Subcontractor' }
        post:
            tags: [subcontractors]
            summary: เพิ่มผู้รับเหมาช่วง (ต้องมีสิทธิ์ project.manage) — หลังบ้านออกรหัส SUB-xxx
            operationId: createSubcontractor
            requestBody:
                required: true
                content:
                    application/json:
                        schema: { $ref: '#/components/schemas/SubcontractorInput' }
            responses:
                '201':
                    description: เพิ่มแล้ว
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/Subcontractor' }
                '403': { $ref: '#/components/responses/Forbidden' }
                '409': { $ref: '#/components/responses/Conflict' }
                '422': { $ref: '#/components/responses/ValidationError' }
    /subcontractors/{id}:
        parameters:
            - { name: id, in: path, required: true, schema: { type: string } }
        put:
            tags: [subcontractors]
            summary: แก้ไขข้อมูลผู้รับเหมาช่วง / เปิด-ปิดใช้งาน
            operationId: updateSubcontractor
            requestBody:
                required: true
                content:
                    application/json:
                        schema: { $ref: '#/components/schemas/SubcontractorInput' }
            responses:
                '200':
                    description: บันทึกแล้ว
                    content:
                        application/json:
                            schema: { $ref: '#/components/schemas/Subcontractor' }
                '403': { $ref: '#/components/responses/Forbidden' }
                '404': { $ref: '#/components/responses/NotFound' }
                '409': { $ref: '#/components/responses/Conflict' }
                '422': { $ref: '#/components/responses/ValidationError' }
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
    parameters:
        ProjectCode:
            name: code
            in: path
            required: true
            schema: { type: string, examples: [CR690001] }
...
    responses:
...
        Forbidden:
            description: บทบาทนี้ไม่มีสิทธิ์
            content:
                application/problem+json:
                    schema: { $ref: '#/components/schemas/Problem' }
        NotFound:
            description: ไม่พบข้อมูล
            content:
                application/problem+json:
                    schema: { $ref: '#/components/schemas/Problem' }
        Conflict:
            description: สถานะข้อมูลเปลี่ยนไปแล้ว เช่น คำขอถูกตัดสินไปก่อนหน้า
            content:
                application/problem+json:
                    schema: { $ref: '#/components/schemas/Problem' }
        ValidationError:
            description: ข้อมูลไม่ถูกต้อง
            content:
                application/problem+json:
                    schema: { $ref: '#/components/schemas/Problem' }
    schemas:
        Problem:
            type: object
            required: [title, status]
            properties:
                type: { type: string, format: uri, default: about:blank }
                title: { type: string, description: ข้อความสั้นสำหรับแสดงผู้ใช้ (ภาษาไทย) }
                status: { type: integer }
                detail: { type: string }
                errors:
                    type: object
                    description: ข้อผิดพลาดรายช่อง (ชื่อช่อง → ข้อความ)
                    additionalProperties: { type: string }
...
        SubcontractorTrade:
            type: string
            description: สาขางานของผู้รับเหมาช่วง (หน้าบ้านแปลเป็นข้อความไทย)
            enum: [piling, structure, masonry, roofing, waterproofing, electrical, plumbing, hvac, aluminium, tiling, ceiling, painting, carpentry, landscape, solar, pool, lift, fire, other]
        SubcontractorInput:
            type: object
            required: [name, trades, contactName, phone]
            properties:
                name: { type: string, minLength: 1, description: ชื่อบริษัท/ทีมช่าง }
                trades: { type: array, minItems: 1, items: { $ref: '#/components/schemas/SubcontractorTrade' } }
                contactName: { type: string, minLength: 1 }
                phone: { type: string, description: ตัวเลข 9-10 หลัก ขึ้นต้นด้วย 0 }
                email: { type: string }
                lineId: { type: string }
                taxId: { type: string, description: เลขประจำตัวผู้เสียภาษี 13 หลัก (ซ้ำกับรายอื่นไม่ได้ 409) }
                address: { type: string }
                note: { type: string }
                status: { type: string, enum: [active, inactive], default: active, description: inactive = ไม่ใช้งานแล้ว มอบหมายงานใหม่ไม่ได้ }
        Subcontractor:
            allOf:
                - $ref: '#/components/schemas/SubcontractorInput'
                - type: object
                  required: [id, code, status, activeProjects]
                  properties:
                      id: { type: string }
                      code: { type: string, examples: [SUB-001] }
                      activeProjects: { type: integer, description: จำนวนโครงการที่ยังไม่ส่งมอบที่มอบหมายอยู่ }
        StaffRole:
            type: string
            enum: [project-manager, engineer, site-supervisor, architect, foreman, safety-officer, purchasing, accounting]
        StaffAssignmentInput:
            type: object
            required: [personnelId, role]
            properties:
                personnelId: { type: string }
                role: { $ref: '#/components/schemas/StaffRole' }
                note: { type: string, description: เช่น หน้าที่เฉพาะในโครงการนี้ }
        StaffAssignment:
            type: object
            required: [id, personnelId, name, role]
            properties:
                id: { type: string }
                personnelId: { type: string }
                name: { type: string }
                position: { type: string }
                phone: { type: string }
                role: { $ref: '#/components/schemas/StaffRole' }
                note: { type: string }
        SubcontractorAssignmentInput:
            type: object
            required: [subcontractorId, scope, phaseCodes]
            properties:
                subcontractorId: { type: string }
                scope: { type: string, minLength: 1, description: ขอบเขตงานที่จ้าง }
                phaseCodes: { type: array, minItems: 1, items: { type: string }, description: ขั้นตอนในไทม์ไลน์ที่รับผิดชอบ }
                contractValue: { type: number, minimum: 0, description: มูลค่าจ้าง (บาท) }
                note: { type: string }
        SubcontractorAssignment:
            type: object
            required: [id, subcontractor, scope, phaseCodes, start, end, progress, status]
            properties:
                id: { type: string }
                subcontractor:
                    type: object
                    required: [id, code, name, contactName, phone, trades]
                    properties:
                        id: { type: string }
                        code: { type: string }
                        name: { type: string }
                        contactName: { type: string }
                        phone: { type: string }
                        trades: { type: array, items: { $ref: '#/components/schemas/SubcontractorTrade' } }
                scope: { type: string }
                phaseCodes: { type: array, items: { type: string } }
                contractValue: { type: number, description: ส่งเฉพาะผู้มีสิทธิ์ project.manage }
                note: { type: string }
                start: { type: string, format: date, description: วันเริ่มของขั้นตอนแรกที่รับผิดชอบ }
                end: { type: string, format: date }
                progress: { type: integer, minimum: 0, maximum: 100, description: ความคืบหน้าของขั้นตอนที่รับผิดชอบ (ถ่วงตามระยะเวลา) }
                status: { type: string, enum: [upcoming, working, done] }
        ProjectAssignments:
            type: object
            required: [staff, subcontractors]
            properties:
                staff: { type: array, items: { $ref: '#/components/schemas/StaffAssignment' } }
                subcontractors: { type: array, items: { $ref: '#/components/schemas/SubcontractorAssignment' } }
...